requirements.txt
tensorboard==2.1.0
tensorflow-gpu==2.1.0
numpy==1.18.1
scipy==1.4.1
scikit-image==0.16.2
colour-demosaicing
rawkit
exifread
tqdm
imageio
matplotlib
tikzplotlib
sewar
dahuffman
sklearn
python-Levenshtein
latex
pdf2image
loguru